[ref] make CPU usage percentage an integer

+++ backend/dashboard/Models/CpuInfo.cs
@@ -2,5 +2,5 @@ namespace dashboard.Models;
 
 public class CpuInfo
 {
-    public double Usage { get; set; }
+    public int Usage { get; set; }
 }
